fix: move tf model loader closer to the service
