+++ apps/main/src/server/common/api-handler.ts
@@ -1,7 +1,3 @@
-// prettier-ignore-start
-import * as toxicity from '@tensorflow-models/toxicity';
-import '@tensorflow/tfjs-node';
-// prettier-ignore-end
 import { NextApiRequest, NextApiResponse } from 'next';
 import connect, { ErrorHandler } from 'next-connect';
 
@@ -27,25 +23,3 @@ export const getApiHandler = () =>
   connect<NextApiRequest, NextApiResponse>({
     onError: handleInternalFailure,
   });
-
-const MIN_PREDICTION_CONFIDENCE = 0.9;
-
-// Load the tfjs models whenever there is a request to the server,
-// to speed up the loading process (the models are pretty big)
-const toxicModelPromise = toxicity.load(MIN_PREDICTION_CONFIDENCE, [
-  'toxicity',
-  'severe_toxicity',
-  `identity_attack`,
-  `insult`,
-  `threat`,
-  `sexual_explicit`,
-  `obscene`,
-]);
-
-export function getToxicModel() {
-  // Save the ref to the global env to avoid reload the models between severless calls
-  if (!global.toxicModelPromise) {
-    global.toxicModelPromise = toxicModelPromise;
-  }
-  return global.toxicModelPromise;
-}
