+++ apps/main/src/server/services/content-classifier/.prettierignore
@@ -1,0 +1,1 @@
+toxic-text.ts
